Refactor Subscription Handling and Admin Subscription relations.
- Modified AdminsController to return current subscription ID.
- Updated SubscriptionsController to include additional subscription details in responses.
- Added StartDate and EndDate to Subscription model and updated migrations.
- Update Admin model to manage subscriptions as collection instead of a single ID.
- Updated AdminErrors to include new error for active subscriptions.
- Refactor CreateSubscriptionCommandHandler to validate admin existence and active subscriptions.
- Refactor DeleteSubscriptionCommand to check if user exists before delete.
- Update SubscriptionResponse to include additional fields Price, MaxRooms, MaxDailySessions and AdminId.
- Update database context and migrations to reflect changes in subscription and admin relationships.

+++ src/GymManagement.Application/Subscriptions/Commands/DeleteSubscription/DeleteSubscriptionCommandHandler.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using ErrorOr;
 using GymManagement.Application.Common.Interfaces;
+using GymManagement.Domain.Admins;
 using GymManagement.Domain.Subscriptions;
 using MediatR;
 
@@ -12,23 +13,29 @@ namespace GymManagement.Application.Subscriptions.Commands.DeleteSubscription;
 public class DeleteSubscriptionCommandHandler : IRequestHandler<DeleteSubscriptionCommand, ErrorOr<Deleted>>
 {
     private readonly ISubscriptionsRepository _subscriptionsRepository;
+    private readonly IAdminsRepository _adminsRepository;
     private readonly IUnitOfWork _unitOfWork;
-    public DeleteSubscriptionCommandHandler(ISubscriptionsRepository subscriptionsRepository,
-                                          IUnitOfWork unitOfWork)
+                                           
+    public DeleteSubscriptionCommandHandler(IUnitOfWork unitOfWork,
+                                            ISubscriptionsRepository subscriptionsRepository,
+                                            IAdminsRepository adminsRepository)
     {
         _subscriptionsRepository = subscriptionsRepository;
         _unitOfWork = unitOfWork;
+        _adminsRepository = adminsRepository;
     }
 
     public async Task<ErrorOr<Deleted>> Handle(DeleteSubscriptionCommand command, CancellationToken cancellationToken = default)
     {
-        var subscription = await _subscriptionsRepository.GetByIdAsync(command.Id, cancellationToken);
+        var subscription = await _subscriptionsRepository.GetByIdAsync(command.SubscriptionId, cancellationToken);
         if (subscription is null)
-            return Error.NotFound(code: "Subscription.NotFound",
-                                  description: $"Subscription with ID '{command.Id}' not found.");
-        
-        // TODO: check if user exist, remove/cancel subscription from user/admins, 
+            return SubscriptionErrors.SubscriptionNotFound(command.SubscriptionId);
 
+        var admin = await _adminsRepository.GetByIdAsync(subscription.AdminId, cancellationToken);
+        if (admin is null)
+            return AdminErrors.UserNotFound(subscription.AdminId);
+
+        // TODO: Check if can allow remove active subscription?        
         await _subscriptionsRepository.RemoveSubscription(subscription, cancellationToken);
         await _unitOfWork.CommitChangesAsync(cancellationToken);
